refactor: Abstract copy to clipboard logic to utils and update components
